# backend/pkg/swagger/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a mall-tenants API",
        "title": "mall-tenants",
        "contact": {
            "name": "Dmitry Boyarkin",
            "email": "boyarkin_dima2@mail.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/premise": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает список помещений",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "premises"
                ],
                "summary": "Получить список помещений",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Premises"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Помещений не найдено",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/premise/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новое помещение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "premises"
                ],
                "summary": "Создать помещение",
                "parameters": [
                    {
                        "description": "Информация о помещении",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Premises"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет данные о помещении",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "premises"
                ],
                "summary": "Обновить помещение",
                "parameters": [
                    {
                        "description": "Информация о помещении",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Premises"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Помещение не найдено",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/rental/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает данные о аренде по коду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rental"
                ],
                "summary": "Получить данные о аренде",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Код аренды",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Rental"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Аренда не найдена",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет данные аренды по коду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rental"
                ],
                "summary": "Обновить аренду",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Код аренды",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные аренды",
                        "name": "rental",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Rental"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Rental"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Аренда не найдена",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Premises": {
            "type": "object",
            "properties": {
                "air_conditioning": {
                    "type": "string"
                },
                "area": {
                    "type": "integer"
                },
                "code": {
                    "type": "integer"
                },
                "floor": {
                    "type": "integer"
                },
                "rent_per_month": {
                    "type": "integer"
                },
                "security_system": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.PremisesType"
                }
            }
        },
        "model.PremisesType": {
            "type": "string",
            "enum": [
                "retail",
                "service"
            ],
            "x-enum-varnames": [
                "Retail",
                "Service"
            ]
        },
...
        "model.Rental": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "paid_months": {
                    "type": "integer"
                },
                "rental_id": {
                    "type": "integer"
                },
                "space_code": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "sys.ErrorResponse": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}